bonus/requirements.txt
faker==20.1.0
psycopg2-binary==2.9.9
cassandra-driver==3.29.0
pymongo==4.6.0
redis==5.0.1
neo4j==5.15.0
